backend/app/requirements.txt
fastapi
uvicorn
requests
google-auth
google-auth-httplib2
google-auth-oauthlib

